refactor: エラーハンドラーを infrastructure レイヤーに移動
