+++ src/renderer/src/infrastructure/error-handler.test.ts
