+++ src/renderer/src/infrastructure/error-handler.ts
