+++ src/renderer/src/main.ts
@@ -2,7 +2,7 @@ import { mount } from 'svelte';
 import './index.css';
 
 import App from './App.svelte';
-import { initializeGlobalErrorHandler } from './utils/error-handler';
+import { initializeGlobalErrorHandler } from './infrastructure/error-handler';
 import { initializePlatform } from './constants/platform';
 
 const init = async (): Promise<void> => {
